requirements.txt
boto3==1.34.162
python-dotenv==1.0.1
pdfplumber==0.11.4
camelot-py[cv]==0.11.0
opencv-python==4.10.0.84
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.4.0
faiss-cpu==1.8.0
numpy==1.26.4
pandas==2.2.2